+++ contexts/PaymentContext/tests/Unit/Domain/PaymentUrlGenerator/PayPalConfigTest.php
@@ -1,0 +1,48 @@
+<?php
+
+declare( strict_types = 1 );
+
+namespace WMDE\Fundraising\Frontend\Tests\Unit\PaymentContext\Domain\PaymentUrlGenerator;
+
+use WMDE\Fundraising\Frontend\PaymentContext\Domain\PaymentUrlGenerator\PayPalConfig;
+
+/**
+ * @covers \WMDE\Fundraising\Frontend\PaymentContext\Domain\PaymentUrlGenerator\PayPalConfig
+ *
+ * @licence GNU GPL v2+
+ * @author Kai Nissen < kai.nissen@wikimedia.de >
+ */
+class PayPalConfigTest extends \PHPUnit\Framework\TestCase {
+
+	public function testGivenIncompletePayPalUrlConfig_exceptionIsThrown() {
+		$this->expectException( \RuntimeException::class );
+		$this->newIncompletePayPalUrlConfig();
+	}
+
+	private function newIncompletePayPalUrlConfig() {
+		return PayPalConfig::newFromConfig( [
+			PayPalConfig::CONFIG_KEY_BASE_URL => 'http://that.paymentprovider.com/?',
+			PayPalConfig::CONFIG_KEY_ACCOUNT_ADDRESS => 'some@email-adress.com',
+			PayPalConfig::CONFIG_KEY_NOTIFY_URL => 'http://my.donation.app/handler/paypal/',
+			PayPalConfig::CONFIG_KEY_RETURN_URL => 'http://my.donation.app/donation/confirm/',
+			PayPalConfig::CONFIG_KEY_CANCEL_URL => 'http://my.donation.app/donation/cancel/',
+			PayPalConfig::CONFIG_KEY_ITEM_NAME => ''
+		] );
+	}
+
+	public function testGivenValidPayPalUrlConfig_payPalConfigIsReturned() {
+		$this->assertInstanceOf( PayPalConfig::class, $this->newPayPalUrlConfig() );
+	}
+
+	private function newPayPalUrlConfig() {
+		return PayPalConfig::newFromConfig( [
+			PayPalConfig::CONFIG_KEY_BASE_URL => 'http://that.paymentprovider.com/?',
+			PayPalConfig::CONFIG_KEY_ACCOUNT_ADDRESS => 'some@email-adress.com',
+			PayPalConfig::CONFIG_KEY_NOTIFY_URL => 'http://my.donation.app/handler/paypal/',
+			PayPalConfig::CONFIG_KEY_RETURN_URL => 'http://my.donation.app/donation/confirm/',
+			PayPalConfig::CONFIG_KEY_CANCEL_URL => 'http://my.donation.app/donation/cancel/',
+			PayPalConfig::CONFIG_KEY_ITEM_NAME => 'This appears on the invoice'
+		] );
+	}
+
+}
